requirements.txt
streamlit
google-generativeai==0.3.1